Move ProgressBar to new file
